Updating web_ui to add API endpoints
This splits up the previous `search` request handler method into an
internal method which does the bulk of the work and two handler methods
that call it. One is the existing endpoint used by the web UI and the
other is a simple HTTP API endpoint which accepts either GET or POST
requests and returns a JSON-ified list of documents.